requirements.txt
matplotlib == 2.0.0
numpy == 1.15.4
pandas == 0.23.4
rpy2 == 3.1.0
scipy == 1.1.0
seaborn == 0.9.0
setuptools == 39.0.1
sklearn == 0.0
tqdm == 4.32.2
trimap == 1.4.3.dev1
umap_learn == 0.3.7
xgboost == 0.81
